+++ studio/src/app/pages/core/app-signin/app-signin.tsx
@@ -268,13 +268,13 @@ export class AppSignIn {
         if (this.redirect === 'editor') {
             return [
                 <h1 class="ion-text-center ion-padding-start ion-padding-end">Oh, hi! Good to have you.</h1>,
-                <p class="ion-text-center ion-padding">Sign in to add more slides, to use your own images, to publish your presentation and to
+                <p class="ion-text-center ion-padding">Sign in to add more slides, to use your own images, to share your presentation and to
                     get soon a personalized feed of recommendations.</p>
             ]
         } else {
             return [
                 <h1 class="ion-text-center ion-padding-start ion-padding-end">Oh, hi! Welcome back.</h1>,
-                <p class="ion-text-center ion-padding">Sign in to publish your presentation and to get soon a
+                <p class="ion-text-center ion-padding">Sign in to share your presentation and to get soon a
                     personalized feed of recommendations.</p>
             ]
         }