+++ studio/src/app/utils/editor/more-action.tsx
@@ -1,0 +1,7 @@
+export enum MoreAction {
+    FULLSCREEN,
+    JUMP_TO,
+    REMOTE,
+    SHARE,
+    PUBLISH
+}