+++ docs/CHANGELOG.md
@@ -1,3 +1,11 @@
+<a name="1.0.0-alpha.20"></a>
+# 1.0.0-alpha.20 (2019-07-13)
+
+### Features
+
+* add new component social option `dev`
+* style menu hover in dark mode
+
 <a name="1.0.0-alpha.19"></a>
 # 1.0.0-alpha.19 (2019-07-11)